LayoutTests/fast/dom/shadow/shadowroot-clonenode.html

<!DOCTYPE>
<html>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/polyfill.js"></script>

<pre id="console"></pre>

<script>
description('Calling ShadowRoot.cloneNode() should throw a DATA_CLONE_ERR exception.');

var host = document.createElement('div');
var shadowRoot = host.webkitCreateShadowRoot();
shouldThrow('shadowRoot.cloneNode()', '"Error: DataCloneError: DOM Exception 25"');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</html>
